+++ src/main/java/models/GetProfileResponse.java
@@ -12,7 +12,7 @@ import java.util.List;
 @NoArgsConstructor
 @Builder
 public class GetProfileResponse extends BaseModel {
-    private int id;
+    private long id;
     private String username;
     private String password;
     private String name;
@@ -24,7 +24,7 @@ public class GetProfileResponse extends BaseModel {
     @NoArgsConstructor
     @Builder
     public static class Account {
-        private int id;
+        private long id;
         private String accountNumber;
         private double balance;
         private List<Transaction> transactions;
@@ -35,10 +35,10 @@ public class GetProfileResponse extends BaseModel {
     @NoArgsConstructor
     @Builder
     public static class Transaction {
-        private int id;
+        private long id;
         private double amount;
         private String type;
         private String timestamp;
-        private int relatedAccountId;
+        private long relatedAccountId;
     }
 }
