+++ src/main/java/models/TransferRequest.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Builder
 public class TransferRequest extends BaseModel {
-    private int senderAccountId;
-    private int receiverAccountId;
+    private long senderAccountId;
+    private long receiverAccountId;
     private double amount;
 }
